requirements.txt
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0
googlemaps==4.10.0
openpyxl==3.0.9
python-dotenv==0.19.0
playwright==1.17.1
requests==2.32.3
setuptools==65.5.1
pip==23.0.1 